fix: assign viewer role when accepting invitation for user (#108)
There are two issues with the current invitation system:
- When the invitation is accepted, we are not assigning any roles to the
new organization user created. We should assign the viewer role.
- We are currently only accepting invitation after the authentication
flow completes, so if I am creating an invitation for an email that is
associated with an existing account, that account will have to log out
and log in again to be added to the organization, which is not ideal.
Instead, when creating invitations, if an account for the email being
already exists, we add a new user for it to the organization.