fix: cancel external resource when cancelling execution (#272)
When cancelling executions, we are not stopping the third-party
resources created as part of executions, which can lead to unwanted
things running.

This pull request updates the API and execution poller workers to handle
this case. The API now only sets the cancelled_at timestamp when an
execution is cancelled, and the worker handles executions with
cancelled_at timestamp set properly, stopping the external resource. In
order to allow for this to happen for any integration we have, the
ResourceManager interface was updated to include a new Cancel() method.

### Other changes

- Removed old and unused STATE_REASON_STUCK and STATE_REASON_TIMEOUT
reasons from stage events.
- Remove execution resource poller and consolidates all logic for
monitoring execution and execution resources in the same worker.