feat: manual event source type (#333)
Introduces manual event source type support.

### Typed event sources

While implementing this new kind of event source, it became that it
would be helpful if the event sources were typed, the same way stage
executors are typed. That simplifies how the UI manages the different
types of event sources, and it also allows us to have have some other
controls in place: manual event sources can only be triggered using the
`POST /events` endpoint - while not being able to use the `POST
/sources/{id}` one.

For that reason, a new `type` column was added to the event_sources
table with migration to populate the existing sources accordingly. That
type is either `manual`, `scheduled`, or `webhook`, or the type of the
integration (github, semaphore, ...)

### API changes

When creating an event source, the `key` is only returned now for
`webhook` event sources.

### UI

- New manual event source in ComponentSidebar
- Refactor how components are defined in ComponentSidebar to use the
actual spec of the component
- Added utility functions for component type handling