chore: Switch base image to Ubuntu to support playwright (#424)
This is a more drastic change so I feel the need to justify it.

- Playwright (cypress also partially) only really support Ubuntu
- Official golang images are based on Debian or Alpine

To support this:

- I started from the official Ubuntu image, the same one we use for the
release
- Manually installed Golang in the base

---

At some point in development, I was struggling to understand why things
are not working, so I extracted long install steps into dedicated files
where I can more easily add debug messages.

If you feel this is not nice for the future, I can inline them back into
the Dockerfile.

Signed-off-by: Igor Šarčević <igor@operately.com>