Refactor/metadata spec status pattern in workflows (#821)
Related to https://github.com/superplanehq/superplane/issues/633

The old system APIs used the metadata/spec/status pattern. We should use
that pattern for the new system's API as well. To kick it off, I'm
starting with the DescribeWorkflow endpoint. Doing so will allow me to
add a `status` field in the Workflow object in a subsequent PR to return
all the last executions / next queue items for all the nodes.