feat: initial pages for app installations + semaphore application (#1334)
### New app installation pages on the organization

New pages for installing applications, and describing app installations
are added to the organization.

### Semaphore application

Implementation of the Application interface for Semaphore. Two initial
blocks added:
- onPipelineDone - same as the current semaphore trigger
- runWorkflow - same as the current semaphore component

### Blocks from application are displayed in the sidebar

All the blocks provided by the available applications are displayed in
the building blocks sidebar.

### App installations are soft-deleted

Since app installations can have webhooks associated to it, we cannot
delete the app installation it before cleaning up the webhooks. To
handle that, app installations are now soft-deleted. Hard deletion is
handled by the new AppInstallationCleanupWorker, that only deletes the
app installation if all webhooks were cleaned up.