fix: merge component uses new payload structure (#1460)
Executions for the merge component are handled mostly as part of
ProcessQueueItem(), and ProcessQueueContext does not provide an
interface that forces the component to emit events with the new base
payload structure. To address that, we:
- Simplify the ProcessQueueContext by returning `core.ExecutionContext`
from methods that get/create executions. With that, the interface for
managing state/metadata/kv for executions is the same across all
methods.
- We should not have a dependency between pkg/core and pkg/models, so
the models.WorkflowNodeExecution references were removed. The only
reason we need them is so the queue worker can send updates to the UI,
and we can achieve through IDs only.

Signed-off-by: Lucas Pinheiro <lucas@superplane.com>